requirements.txt
beautifulsoup4==4.13.3
google_api_python_client==2.162.0
google_api_python_client==2.163.0
google_auth_oauthlib==1.2.1
Jinja2==3.1.6
openai==1.66.3
protobuf==6.30.1
python-dotenv==1.0.1
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
SQLAlchemy==2.0.38
